Pull out FailedJobsException to break the import loop
