+++ src/toil/test/lib/dockerTest.py
@@ -21,7 +21,7 @@ from threading import Thread
 from docker.errors import ContainerError
 from toil.common import Toil
 from toil.job import Job
-from toil.leader import FailedJobsException
+from toil.exceptions import FailedJobsException
 from toil.lib.docker import (FORGO,
                              RM,
                              STOP,
