+++ src/toil/test/src/checkpointTest.py
@@ -14,7 +14,7 @@
 
 from toil.job import Job
 from toil.jobStores.abstractJobStore import NoSuchFileException
-from toil.leader import FailedJobsException
+from toil.exceptions import FailedJobsException
 from toil.test import ToilTest, slow
