+++ src/toil/test/src/importExportFileTest.py
@@ -20,7 +20,7 @@ import uuid
 from toil.common import Toil
 from toil.fileStores import FileID
 from toil.job import Job
-from toil.leader import FailedJobsException
+from toil.exceptions import FailedJobsException
 from toil.test import ToilTest, slow
