+++ src/toil/test/src/jobServiceTest.py
@@ -25,7 +25,8 @@ import pytest
 
 from toil.batchSystems.singleMachine import SingleMachineBatchSystem
 from toil.job import Job
-from toil.leader import DeadlockException, FailedJobsException
+from toil.leader import DeadlockException
+from toil.exceptions import FailedJobsException
 from toil.lib.retry import retry_flaky_test
 from toil.test import ToilTest, get_temp_file, slow
