+++ src/toil/test/src/resumabilityTest.py
@@ -16,7 +16,7 @@ import os
 
 from toil.job import Job
 from toil.jobStores.abstractJobStore import NoSuchFileException
-from toil.leader import FailedJobsException
+from toil.exceptions import FailedJobsException
 from toil.test import ToilTest, slow
