+++ src/toil/test/src/retainTempDirTest.py
@@ -15,7 +15,7 @@ import os
 import shutil
 
 from toil.job import Job
-from toil.leader import FailedJobsException
+from toil.exceptions import FailedJobsException
 from toil.test import ToilTest
